Split constants into 32 and 64 bit to satisfy x86 constraints.
The only allowable imm64 constant in x86 operations is a quad register move.
Every other constant must be treated as imm32. CST has been split into CSTL and
CSTQ with some special rules to enforce this.

+++ test-utils.c
@@ -41,7 +41,7 @@ Value *l(int label) {
     return v;
 }
 
-Value *c(int value) {
+Value *c(long value) {
     return new_constant(TYPE_LONG, value);
 }
